Fix type of reads in kb

+++ cg/services/pacbio/metrics/models.py
@@ -14,7 +14,7 @@ class HiFiMetrics(BaseModel):
 
     reads: int = Field(..., alias=CCSAttributeIDs.NUMBER_OF_READS)
     yield_: int = Field(..., alias=CCSAttributeIDs.TOTAL_NUMBER_OF_BASES)
-    mean_read_length_kb: int = Field(..., alias=CCSAttributeIDs.MEAN_READ_LENGTH)
+    mean_read_length_kb: float = Field(..., alias=CCSAttributeIDs.MEAN_READ_LENGTH)
     median_read_length: int = Field(..., alias=CCSAttributeIDs.MEDIAN_READ_LENGTH)
     mean_length_n50: int = Field(..., alias=CCSAttributeIDs.READ_LENGTH_N50)
     median_read_quality: str = Field(..., alias=CCSAttributeIDs.MEDIAN_ACCURACY)
@@ -30,7 +30,7 @@ class ControlMetrics(BaseModel):
     """Model for the control metrics."""
 
     reads: int = Field(..., alias=ControlAttributeIDs.NUMBER_OF_READS)
-    mean_read_length_kb: int = Field(..., alias=ControlAttributeIDs.MEAN_READ_LENGTH)
+    mean_read_length_kb: float = Field(..., alias=ControlAttributeIDs.MEAN_READ_LENGTH)
     percent_mean_concordance_reads: float = Field(
         ..., alias=ControlAttributeIDs.PERCENT_MEAN_READ_CONCORDANCE
     )
